Lots of bootstrap packages

+++ tools/package_builder.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3.10
 
 import argparse
+import os
 from pathlib import Path
 import shutil
 import subprocess
-import tempfile
 
 WORKSPACE = 'workspace'
 
@@ -34,9 +34,7 @@ def parse_dependencies(package_dir):
 
 
 class Repository:
-    # We expect that if a package is compiled then
-    # all of its dependencies are also compiled
-    def __init__(self, package_dir):
+    def __init__(self, package_dir, ld_path=Path('tools/osoap-ld').resolve()):
         self.package_dir = package_dir.resolve()
         self.dependencies = parse_dependencies(package_dir)
         self.reverse_dependencies = {}
@@ -45,6 +43,12 @@ class Repository:
         for name, deps in self.dependencies.items():
             for dep in deps:
                 self.reverse_dependencies[dep].add(name)
+
+        self.compile_env = os.environ.copy()
+        assert ld_path.exists() and ld_path.is_absolute(), ld_path
+        self.compile_env['LD'] = str(ld_path)
+
+        self.sysroot = self.package_dir / 'sysroot'
 
     def validate(self):
         current_status = {}
@@ -59,28 +63,29 @@ class Repository:
 
     @property
     def packages(self):
-        return self.dependencies.keys()
+        return sorted(self.dependencies.keys())
 
-    def clean(self, package):
+    def clean(self, package, rev_recursive=False):
         package_workspace = self.package_dir / package / WORKSPACE
         if not package_workspace.is_dir():
             return False
-        for rev_dep in self.reverse_dependencies[package]:
-            self.clean(rev_dep)
+        if rev_recursive:
+            for rev_dep in self.reverse_dependencies[package]:
+                self.clean(rev_dep, True)
         shutil.rmtree(package_workspace)
         return True
 
-    def _install(self, package, sysroot, seen):
+    def _install(self, package, seen):
         if package in seen:
             return
         for dep in self.dependencies[package]:
-            self._install(dep, sysroot, seen)
+            self._install(dep, seen)
         package_workspace = self.package_dir / package / WORKSPACE
         assert package_workspace.is_dir()
         subprocess.run(
             [
                 self.package_dir / package / 'install',
-                sysroot,
+                self.sysroot,
                 package_workspace,
             ],
             cwd=self.package_dir / package,
@@ -89,48 +94,49 @@ class Repository:
 
     def build(self, packages):
         """Must ensure all packages are compiled beforehand."""
-        sysroot = None
         try:
-            sysroot = tempfile.TemporaryDirectory(
-                    prefix='sysroot', ignore_cleanup_errors=True)
-            seen = set()
-            for package in packages:
-                self._install(package, sysroot, seen)
-        except BaseException:
-            if sysroot is not None:
-                sysroot.cleanup()
-            raise
-        return sysroot
+            shutil.rmtree(self.sysroot)
+        except FileNotFoundError:
+            pass
 
-    def compile(self, package):
+        self.sysroot.mkdir()
+
+        seen = set()
+        for package in packages:
+            self._install(package, seen)
+
+        return self.sysroot
+
+    def compile(self, package, cleanup_on_failure=True):
         package_workspace = self.package_dir / package / WORKSPACE
-        if package_workspace.is_dir():
-            return False
         for dep in self.dependencies[package]:
             self.compile(dep)
+        if package_workspace.is_dir():
+            return False
 
-        with self.build(self.dependencies[package]) as sysroot:
-            try:
-                subprocess.run(
-                    [
-                        self.package_dir / package / 'compile',
-                        sysroot,
-                        package_workspace,
-                    ],
-                    cwd=self.package_dir / package,
-                    check=True)
-            except BaseException:
-                if package_workspace.exists():
-                    shutil.rmtree(package_workspace)
-                raise
-            assert package_workspace.is_dir()
+        self.build(self.dependencies[package])
+        try:
+            subprocess.run(
+                [
+                    self.package_dir / package / 'compile',
+                    self.sysroot,
+                    package_workspace,
+                ],
+                cwd=self.package_dir / package,
+                env=self.compile_env,
+                check=True)
+        except BaseException:
+            if cleanup_on_failure and package_workspace.exists():
+                shutil.rmtree(package_workspace)
+            raise
+        assert package_workspace.is_dir()
         return True
 
 
-def friendly_clean(pkg, repository):
+def friendly_clean(pkg, repository, rev_recursive=False):
     if pkg in repository.packages:
         print(f"Cleaning {pkg}...")
-        if repository.clean(pkg):
+        if repository.clean(pkg, rev_recursive):
             print(f"{pkg} has been cleaned.")
         else:
             print(f"{pkg} was already clean.")
@@ -138,10 +144,10 @@ def friendly_clean(pkg, repository):
         print(f"Package {pkg} not recognized, skipping.")
 
 
-def friendly_compile(pkg, repository):
+def friendly_compile(pkg, repository, cleanup_on_failure=True):
     if pkg in repository.packages:
         print(f"Compiling {pkg}...")
-        if repository.compile(pkg):
+        if repository.compile(pkg, cleanup_on_failure):
             print(f"{pkg} has been compiled.")
         else:
             print(f"{pkg} was already compiled.")
@@ -153,7 +159,7 @@ def main_clean(args, repository):
     if 'all' in args.package:
         args.package = list(repository.packages)
     for pkg in args.package:
-        friendly_clean(pkg, repository)
+        friendly_clean(pkg, repository, args.rev_recursive)
 
 
 def main_compile(args, repository):
@@ -163,7 +169,7 @@ def main_compile(args, repository):
         for pkg in args.package:
             friendly_clean(pkg, repository)
     for pkg in args.package:
-        friendly_compile(pkg, repository)
+        friendly_compile(pkg, repository, not args.nocleanup)
 
 
 def main_build(args, repository):
@@ -172,6 +178,8 @@ def main_build(args, repository):
     for pkg in args.package:
         friendly_compile(pkg, repository)
 
+    repository.build(args.package)
+
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
@@ -179,37 +187,45 @@ if __name__ == '__main__':
 
     parser.add_argument(
         '--repository',
-        default=Path('bootstrap_packages'),
+        default='bootstrap_packages',
         type=Path,
         help="Defaults to 'bootstrap_packages'",
+    )
+    parser.add_argument(
+        '--ld-path',
+        dest='ld_path',
+        default='tools/osoap-ld',
+        type=Path,
+        help="Defaults to 'tools/osoap-ld'",
     )
 
     subparsers = parser.add_subparsers(required=True)
 
     parser_clean = subparsers.add_parser(
         'clean', help="Remove a compiled package")
+    parser_clean.add_argument(
+        '--rev_recursive', action='store_true',
+        help="Clean packages that depend on this one too")
     parser_clean.add_argument('package', nargs='+')
     parser_clean.set_defaults(func=main_clean)
 
     parser_compile = subparsers.add_parser('compile', help="Compile a package")
     parser_compile.add_argument('--recompile', action='store_true',
                                 help="Recompile even if already compiled")
+    parser_compile.add_argument(
+        '--nocleanup', action='store_true',
+        help="<Advanced> Skip cleaning up the workspace on failure")
     parser_compile.add_argument('package', nargs='+')
     parser_compile.set_defaults(func=main_compile)
 
     parser_build = subparsers.add_parser(
         'build', help="Build a collection of packages")
-    parser_build.add_argument(
-        '--format', choices={'targz', 'dir'}, default='dir',
-        help="What format the resulting collection should be output in.")
-    parser_build.add_argument('destination', type=Path,
-                              help="Where to put the result")
     parser_build.add_argument('package', nargs='*')
     parser_build.set_defaults(func=main_build)
 
     args = parser.parse_args()
 
-    repository = Repository(args.repository)
+    repository = Repository(args.repository, args.ld_path.resolve())
     repository.validate()
 
     try:
